tech(api): move controller method and test to user bounded context
